requirements.txt
BeautifulSoup==3.2.1
dnspython==1.12.0
xmpppy==0.5.0rc1
